+++ rosbag2_transport/test/rosbag2_transport/test_record_options.cpp
@@ -58,6 +58,7 @@ TEST(record_options, test_yaml_serialization_deserialization)
   original.statistics_max_publishing_rate = 5.0f;
   original.repeat_transient_local_messages = {{"/map", 1}, {"/tf_static", 5}};
   original.repeat_all_transient_local_depth = 3;
+  original.topic_throttle_frequencies = {{"/imu", 10.0}, {"/camera/image_raw", 2.5}};
 
   auto node = YAML::convert<rosbag2_transport::RecordOptions>::encode(original);
 
@@ -101,6 +102,7 @@ TEST(record_options, test_yaml_serialization_deserialization)
   CHECK(disable_keyboard_controls);
   CHECK(repeat_transient_local_messages);
   CHECK(repeat_all_transient_local_depth);
+  CHECK(topic_throttle_frequencies);
   #undef CHECK
   ASSERT_FLOAT_EQ(original.statistics_max_publishing_rate,
                   reconstructed.statistics_max_publishing_rate);